+++ src/commands/moderation/nickname.js
@@ -1,19 +1,49 @@
 function nickname(client, evt, suffix) {
   if (evt.message.channel.isPrivate) return evt.message.channel.sendMessage('', false, {color: 3901635, description: `\u2139  Use this command in a server!`});
 
-  if (!evt.message.author.permissionsFor(evt.message.guild).General.MANAGE_NICKNAMES) {
-    let embed = { color: 16763981, description: `\u26A0  You do not have permissions to nickname other members!` };
-    return evt.message.channel.sendMessage('', false, embed);
+  let self = false;
+  let newNick;
+
+  if (evt.message.mentions.length === 0) {
+    self = true;
+    newNick = suffix;
   }
 
-  let suffixArray = suffix.split(' ');
-  if (suffixArray.length < 2) {
-    let embed = { color: 16763981, description: `\u26A0  Something went wrong. Make sure you use the command correctly!` };
-    return evt.message.channel.sendMessage('', false, embed);
+  if (evt.message.mentions.length === 1 && evt.message.mentions[0].id === evt.message.author.id) {
+    self = true;
+    let suffixArray = suffix.split(' ');
+    suffixArray.shift();
+    newNick = suffixArray.join(' ');
   }
-  let newNick = suffixArray[1];
 
-  if (evt.message.mentions.length === 1) {
+  if (evt.message.mentions.length === 1 && evt.message.mentions[0].id !== evt.message.author.id) {
+    let suffixArray = suffix.split(' ');
+    suffixArray.shift();
+    newNick = suffixArray.join(' ');
+  }
+
+  if (self) {
+    if (!evt.message.author.permissionsFor(evt.message.guild).General.CHANGE_NICKNAME) {
+      let embed = { color: 16763981, description: `\u26A0  You do not have permissions to change your nickname on this server!` };
+      return evt.message.channel.sendMessage('', false, embed);
+    }
+    let user = evt.message.author.memberOf(evt.message.guild);
+    user.setNickname(newNick).then(() => {
+      let embed = { color: 7844437, description: `\u2705  Set a new nickname for ${user.username}#${user.discriminator}!` };
+      return evt.message.channel.sendMessage('', false, embed);
+    })
+    .catch(err => {
+      let error = JSON.parse(err.response.error.text);
+      let embed = { color: 16763981, description: `\u26A0  Something went wrong: ${error.message}` };
+      return evt.message.channel.sendMessage('', false, embed);
+    });
+  }
+
+  if (!self) {
+    if (!evt.message.author.permissionsFor(evt.message.guild).General.MANAGE_NICKNAMES) {
+      let embed = { color: 16763981, description: `\u26A0  You do not have permissions to nickname other members!` };
+      return evt.message.channel.sendMessage('', false, embed);
+    }
     let user = evt.message.mentions[0].memberOf(evt.message.guild);
 
     user.setNickname(newNick).then(() => {
@@ -34,5 +64,7 @@ export default {
 };
 
 export const help = {
-  nickname: { parameters: '@User' }
+  nickname: {
+    parameters: ['nickname', 'or', '@User nickname']
+  }
 };